moved chatbot_screen to upper level directory
